backend/requirements.txt
flask>=3.0
python-dotenv>=1.0
flask-cors>=4.0
flasgger>=0.9.7
supabase>=2.0

pytest>=8.0
